add order dialog moved to a separate file, confirm button enabled only when order nam is not empty
